feat: Add filterListLeague method to LeagueController
This commit adds the `filterListLeague` method to the `LeagueController` class. The method is responsible for filtering and returning the first(s) of a league based on their score in descending order. Currently, the implementation is incomplete and marked with a TODO comment.

The purpose of this change is to provide a way to retrieve the top players in a league.